Internal structure refactor
