+++ packages/crypto/src/algorithms-api/index.ts
@@ -1,3 +1,4 @@
 export * from './crypto-algorithm.js';
+export * from './crypto-key.js';
 export * from './ecc/index.js';
 export * from './errors.js';
