+++ packages/crypto/src/crypto-algorithms/index.ts
@@ -1,3 +1,3 @@
-export * from './ed25519.js';
-export * from './secp256k1.js';
-export * from './x25519.js';
+export * from './ecdh.js';
+export * from './ecdsa.js';
+export * from './eddsa.js';
