+++ packages/crypto/src/crypto-primitives/ed25519.ts
