+++ packages/crypto/src/crypto-primitives/index.ts
@@ -1,0 +1,3 @@
+export * from './ed25519.js';
+export * from './secp256k1.js';
+export * from './x25519.js';
