+++ packages/crypto/src/kms-local/index.ts
@@ -1,0 +1,2 @@
+export * from './key-stores.js';
+export * from './kms-local.js';
