--- packages/crypto/src/kms/default/algorithms/index.ts
@@ -1,3 +1,0 @@
-export * from './ecdh.js';
-export * from './ecdsa.js';
-export * from './eddsa.js';
