--- packages/crypto/src/kms/default/index.ts
@@ -1,4 +1,0 @@
-export * from './algorithms/index.js';
-export * from './crypto-key.js';
-export * from './key-stores.js';
-export * from './kms-default.js';
